Move public dir, set up db seed data, etc.
I also updated `compose.yaml` to source `.env` from the `backend`
directory, since that is now the root of the module that needs it
(otherwise `dotenv` won't find my environment variables).

The problem is that `.env` now needs to be in two places at once: the
project root (where `compose.yaml` is) and the `backend` directory. If
I'm always using Docker, I can place it in the root directory and it can
be sourced by all my containers. But since there may be times when I
can't use docker, this setup can work for now.