Update database structure and seed data
For the `authors` table, we'll use an author's Open Library ID (OLID) as
the primary key.

I also added a constraint to the `rating` field to limit the values
between 0 and 5 (inclusive). If I decide to switch to a 10-rating system
in the future, then I can change this constraint with `ALTER TABLE`.